+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Hardened `tdx.converters.TeamConverter` and changed the interface of the return
 - Fixed DivideByZero bug (#11)
 - BaseCard it's descendants are now async - call `await ProfileCard(**kwargs)` instead of `await ProfileCard(**kwargs).build_card(**kwargs)`
+- Fixed `[p]profile create` loop lock issue (#12)
+- Handled errors when we're unable to edit a `discord.Member`
 
 ## [2020.30] - 2020-07-25
 ### Added